Rearrange queries and mutation for auto entries modal
