+++ app/views/EntryEdit/LeftPane/AutoEntriesModal/index.tsx
@@ -1,5 +1,6 @@
 import React, {
     useMemo,
+    useEffect,
     useCallback,
     useState,
 } from 'react';
@@ -10,10 +11,17 @@ import {
     noOp,
     listToMap,
 } from '@togglecorp/fujs';
-import { gql, useQuery } from '@apollo/client';
+import {
+    gql,
+    useQuery,
+    useMutation,
+} from '@apollo/client';
 import {
     Modal,
     ListView,
+    useAlert,
+    Button,
+    Message,
 } from '@the-deep/deep-ui';
 
 import { type Framework } from '#components/entry/types';
@@ -37,6 +45,10 @@ import {
 import {
     AutoEntriesForLeadQuery,
     AutoEntriesForLeadQueryVariables,
+    CreateAutoDraftEntriesMutation,
+    CreateAutoDraftEntriesMutationVariables,
+    AutoDraftEntriesStatusQuery,
+    AutoDraftEntriesStatusQueryVariables,
 } from '#generated/types';
 import AssistPopup from '../AssistItem/AssistPopup';
 
@@ -58,6 +70,7 @@ const AUTO_ENTRIES_FOR_LEAD = gql`
         $leadIds: [ID!],
     ) {
         project(id: $projectId) {
+            id
             assistedTagging {
                 draftEntryByLeads(
                 filter: {
@@ -82,6 +95,39 @@ const AUTO_ENTRIES_FOR_LEAD = gql`
                         threshold
                         value
                     }
+                }
+            }
+        }
+    }
+`;
+
+const CREATE_AUTO_DRAFT_ENTRIES = gql`
+    mutation CreateAutoDraftEntries (
+        $projectId: ID!,
+        $leadId: ID!,
+    ) {
+        project(id: $projectId) {
+            id
+            assistedTagging {
+                autoDraftEntryCreate(data: {lead: $leadId}) {
+                    ok
+                    errors
+                }
+            }
+        }
+    }
+`;
+
+const AUTO_DRAFT_ENTRIES_STATUS = gql`
+    query AutoDraftEntriesStatus (
+        $projectId: ID!,
+        $leadId: ID!,
+    ) {
+        project(id: $projectId) {
+            id
+            assistedTagging {
+                extractionStatusByLead(leadId: $leadId) {
+                    autoEntryExtractionStatus
                 }
             }
         }
@@ -289,15 +335,124 @@ function AutoEntriesModal(props: Props) {
         frameworkDetails,
     } = props;
 
+    const alert = useAlert();
+
     const [
         geoAreaOptions,
         setGeoAreaOptions,
-    ] = useState<{
-        entryId: string;
-        geoAreas: GeoArea[] | undefined | null;
-    } | undefined>();
+    ] = useState<Record<string, GeoArea[] | undefined> | undefined>(undefined);
 
-    // const [messageText, setMessageText] = useState<string | undefined>();
+    // FIXME: randomId is used to create different query variables after each poll
+    // so that apollo doesn't create unnecessary cache
+    const [randomId, setRandomId] = useState<string>(randomString());
+
+    const autoEntryStatusVariables = useMemo(() => {
+        if (isNotDefined(projectId)) {
+            return undefined;
+        }
+        return ({
+            leadId,
+            randomId,
+            projectId,
+        });
+    }, [
+        randomId,
+        leadId,
+        projectId,
+    ]);
+
+    const [draftEntriesLoading, setDraftEntriesLoading] = useState<boolean>(true);
+
+    const {
+        data: autoEntryExtractionStatus,
+        refetch: retriggerEntryExtractionStatus,
+    } = useQuery<AutoDraftEntriesStatusQuery, AutoDraftEntriesStatusQueryVariables>(
+        AUTO_DRAFT_ENTRIES_STATUS,
+        {
+            skip: isNotDefined(autoEntryStatusVariables),
+            variables: autoEntryStatusVariables,
+            onCompleted: (response) => {
+                const status = response?.project
+                    ?.assistedTagging?.extractionStatusByLead?.autoEntryExtractionStatus;
+                if (status === 'SUCCESS') {
+                    setDraftEntriesLoading(false);
+                }
+            },
+        },
+    );
+
+    const extractionStatus = autoEntryExtractionStatus?.project
+        ?.assistedTagging?.extractionStatusByLead?.autoEntryExtractionStatus;
+
+    // TODO: This polling calls two queries at a time. Fix this.
+    useEffect(() => {
+        const timeout = setTimeout(
+            () => {
+                const shouldPoll = autoEntryExtractionStatus?.project
+                    ?.assistedTagging?.extractionStatusByLead?.autoEntryExtractionStatus === 'PENDING';
+                if (shouldPoll) {
+                    setDraftEntriesLoading(true);
+                    setRandomId(randomString());
+                    retriggerEntryExtractionStatus();
+                } else {
+                    setDraftEntriesLoading(false);
+                }
+            },
+            2000,
+        );
+
+        return () => {
+            clearTimeout(timeout);
+        };
+    }, [
+        autoEntryExtractionStatus?.project?.assistedTagging
+            ?.extractionStatusByLead?.autoEntryExtractionStatus,
+        leadId,
+        retriggerEntryExtractionStatus,
+    ]);
+
+    const [
+        triggerAutoEntriesCreate,
+    ] = useMutation<CreateAutoDraftEntriesMutation, CreateAutoDraftEntriesMutationVariables>(
+        CREATE_AUTO_DRAFT_ENTRIES,
+        {
+            onCompleted: (response) => {
+                const autoEntriesResponse = response?.project?.assistedTagging
+                    ?.autoDraftEntryCreate;
+                if (autoEntriesResponse?.ok) {
+                    retriggerEntryExtractionStatus();
+                } else {
+                    alert.show(
+                        'Failed to extract entries using NLP.',
+                        {
+                            variant: 'error',
+                        },
+                    );
+                }
+            },
+            onError: () => {
+                alert.show(
+                    'Failed to extract entries using NLP.',
+                    {
+                        variant: 'error',
+                    },
+                );
+            },
+        },
+    );
+
+    const handleAutoExtractClick = useCallback(() => {
+        triggerAutoEntriesCreate({
+            variables: {
+                projectId,
+                leadId,
+            },
+        });
+    }, [
+        projectId,
+        leadId,
+        triggerAutoEntriesCreate,
+    ]);
 
     const [
         allRecommendations,
@@ -311,7 +466,6 @@ function AutoEntriesModal(props: Props) {
     const mappings = frameworkDetails?.predictionTagsMapping;
 
     const {
-        allWidgets,
         filteredWidgets,
     } = useMemo(() => {
         const widgetsFromPrimary = frameworkDetails?.primaryTagging?.flatMap(
@@ -329,182 +483,6 @@ function AutoEntriesModal(props: Props) {
     }, [
         frameworkDetails,
     ]);
-
-    /*
-    const handleMappingsFetchTest = useCallback(
-        (predictions: { tags: string[]; locations: GeoArea[]; }) => {
-            if (predictions.tags.length <= 0 && predictions.locations.length <= 0) {
-                setMessageText('DEEP could not provide any recommendations for the selected text.');
-                return;
-            }
-
-            setGeoAreaOptions(predictions.locations);
-
-            const matchedMappings = mappings
-                ?.filter(isCategoricalMappings)
-                .filter((m) => m.tag && predictions.tags.includes(m.tag));
-
-            const supportedGeoWidgets = mappings
-                ?.filter((mappingItem) => mappingItem.widgetType === 'GEO')
-                ?.map((mappingItem) => mappingItem.widget);
-
-            const {
-                tempAttrs: recommendedAttributes,
-                tempHints: widgetsHints,
-            } = filteredWidgets.reduce(
-                (
-                    acc: { tempAttrs: PartialAttributeType[]; tempHints: WidgetHint[]; },
-                    widget,
-                ) => {
-                    const {
-                        tempAttrs: oldTempAttrs,
-                        tempHints: oldTempHints,
-                    } = acc;
-
-                    if (widget.widgetId === 'MATRIX1D') {
-                        const supportedTags = matchedMappings
-                            ?.filter((m) => m.widget === widget.id)
-                            .filter(filterMatrix1dMappings);
-
-                        const attr = createMatrix1dAttr(supportedTags, widget);
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: oldTempHints,
-                        };
-                    }
-
-                    if (widget.widgetId === 'MATRIX2D') {
-                        const supportedTags = matchedMappings
-                            ?.filter((m) => m.widget === widget.id)
-                            .filter(filterMatrix2dMappings);
-
-                        const attr = createMatrix2dAttr(supportedTags, widget);
-
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: oldTempHints,
-                        };
-                    }
-                    if (widget.widgetId === 'SCALE') {
-                        const supportedTags = matchedMappings
-                            ?.filter((m) => m.widget === widget.id)
-                            .filter(filterScaleMappings);
-
-                        const {
-                            attr,
-                            hints,
-                        } = createScaleAttr(supportedTags, widget);
-
-                        const hintsWithInfo: WidgetHint | undefined = hints ? {
-                            widgetPk: widget.id,
-                            widgetType: 'SCALE',
-                            hints,
-                        } : undefined;
-
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: hintsWithInfo
-                                ? [...oldTempHints, hintsWithInfo]
-                                : oldTempHints,
-                        };
-                    }
-                    if (widget.widgetId === 'SELECT') {
-                        const supportedTags = matchedMappings
-                            ?.filter((m) => m.widget === widget.id)
-                            .filter(filterSelectMappings);
-
-                        const {
-                            attr,
-                            hints,
-                        } = createSelectAttr(supportedTags, widget);
-
-                        const hintsWithInfo: WidgetHint | undefined = hints ? {
-                            widgetPk: widget.id,
-                            widgetType: 'SELECT',
-                            hints,
-                        } : undefined;
-
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: hintsWithInfo
-                                ? [...oldTempHints, hintsWithInfo]
-                                : oldTempHints,
-                        };
-                    }
-                    if (widget.widgetId === 'MULTISELECT') {
-                        const supportedTags = matchedMappings
-                            ?.filter((m) => m.widget === widget.id)
-                            .filter(filterMultiSelectMappings);
-
-                        const attr = createMultiSelectAttr(
-                            supportedTags,
-                            widget,
-                        );
-
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: oldTempHints,
-                        };
-                    }
-                    if (widget.widgetId === 'ORGANIGRAM') {
-                        const supportedTags = matchedMappings
-                            ?.filter((m) => m.widget === widget.id)
-                            .filter(filterOrganigramMappings);
-
-                        const attr = createOrganigramAttr(
-                            supportedTags,
-                            widget,
-                        );
-
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: oldTempHints,
-                        };
-                    }
-                    if (
-                        widget.widgetId === 'GEO'
-                    && predictions.locations.length > 0
-                    && supportedGeoWidgets?.includes(widget.id)
-                    ) {
-                        const attr = createGeoAttr(
-                            predictions.locations,
-                            widget,
-                        );
-
-                        return {
-                            tempAttrs: attr ? [...oldTempAttrs, attr] : oldTempAttrs,
-                            tempHints: oldTempHints,
-                        };
-                    }
-                    return acc;
-                },
-                {
-                    tempAttrs: [],
-                    tempHints: [],
-                },
-            );
-
-            if (recommendedAttributes.length <= 0 && widgetsHints.length <= 0) {
-                setMessageText(
-                    'The provided recommendations for this text did
-                    not fit any tags in this project.',
-                );
-                return;
-            }
-            setAllHints(widgetsHints);
-            setAllRecommendations(recommendedAttributes);
-            return {
-                hints: widgetsHints,
-                recommendations: recommendedAttributes,
-
-            };
-        },
-        [
-            mappings,
-            filteredWidgets,
-        ],
-    );
-    */
 
     const [
         draftEntries,
@@ -524,7 +502,9 @@ function AutoEntriesModal(props: Props) {
     } = useQuery<AutoEntriesForLeadQuery, AutoEntriesForLeadQueryVariables>(
         AUTO_ENTRIES_FOR_LEAD,
         {
-            skip: isNotDefined(autoEntriesVariables),
+            skip: isNotDefined(extractionStatus)
+                || extractionStatus !== 'SUCCESS'
+                || isNotDefined(autoEntriesVariables),
             variables: autoEntriesVariables,
             onCompleted: (response) => {
                 const entries = response.project?.assistedTagging?.draftEntryByLeads;
@@ -616,7 +596,7 @@ function AutoEntriesModal(props: Props) {
         leadId,
         hints: allHints?.[entryId],
         recommendations: allRecommendations?.[entryId],
-        geoAreaOptions: undefined,
+        geoAreaOptions: geoAreaOptions?.[entryId],
         onEntryDiscardButtonClick: noOp,
         onEntryCreateButtonClick: noOp,
         onNormalEntryCreateButtonClick: noOp,
@@ -631,23 +611,45 @@ function AutoEntriesModal(props: Props) {
         allRecommendations,
         frameworkDetails,
         leadId,
+        geoAreaOptions,
     ]);
+
+    const hideList = draftEntriesLoading
+        || extractionStatus === 'NONE';
 
     return (
         <Modal
             onCloseButtonClick={onModalClose}
-            heading="Draft Entries"
-            size="cover"
+            heading="NLP Extract & Classify"
+            size={hideList ? 'small' : 'cover'}
+            freeHeight={hideList}
         >
             <ListView
                 data={draftEntries}
                 keySelector={entryKeySelector}
                 renderer={AssistPopup}
                 rendererParams={rendererParams}
-                pendingMessage="Please wait while DEEP analyzes your text"
-                pending={autoEntriesLoading}
+                pendingMessage="Please wait while we load recommendations."
+                pending={autoEntriesLoading || draftEntriesLoading}
                 errored={false}
                 filtered={false}
+                emptyMessage={(extractionStatus === 'NONE') && "Looks like you've not triggered an extraction yet"}
+                messageActions={(extractionStatus === 'NONE') && (
+                    <Button
+                        name={undefined}
+                        onClick={handleAutoExtractClick}
+                        variant="tertiary"
+                    >
+                        Recommend entries
+                    </Button>
+                )}
+                messageShown
+                messageIconShown
+                borderBetweenItem
+            />
+            <Message
+                pending={draftEntriesLoading}
+                message=""
             />
         </Modal>
     );
